+++ WinQuake/pr_cmds.c
@@ -447,7 +447,7 @@ void PF_vectoangles (void)
 		if (yaw < 0)
 			yaw += 360;
 
-		forward = sqrt (value1[0]*value1[0] + value1[1]*value1[1]);
+		forward = sqrtf(value1[0]*value1[0] + value1[1]*value1[1]);
 		pitch = (int) (atan2f(value1[2], forward) * 180.f / M_PI);
 		if (pitch < 0)
 			pitch += 360;
